# backend/docs/swagger.yaml
basePath: /
definitions:
  github_com_kohwg_gowoopi_backend_internal_model.AdminLoginRequest:
    properties:
      password:
        type: string
      store_id:
        type: string
      username:
        type: string
    required:
    - password
    - store_id
    - username
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.AuthResponse:
    properties:
      role:
        type: string
      store_id:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.Category:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      sortOrder:
        type: integer
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.CustomerLoginRequest:
    properties:
      password:
        type: string
      store_id:
        type: string
      table_number:
        type: integer
    required:
    - password
    - store_id
    - table_number
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorDetail'
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.Menu:
    properties:
      category:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Category'
      categoryID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      imageURL:
        type: string
      isAvailable:
        type: boolean
      name:
        type: string
      price:
        type: integer
      sortOrder:
        type: integer
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.MenuCreateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      is_available:
        type: boolean
      name:
        type: string
      price:
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.MenuOrderRequest:
    properties:
      id:
        type: integer
      sort_order:
        type: integer
    required:
    - id
    - sort_order
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.MenuUpdateRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      is_available:
        type: boolean
      name:
        type: string
      price:
        type: integer
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.Order:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.OrderItem'
        type: array
      session:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.TableSession'
      sessionID:
        type: string
      status:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.OrderStatus'
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      table:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Table'
      tableID:
        type: integer
      totalAmount:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.OrderCreateRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.OrderItemRequest'
        minItems: 1
        type: array
    required:
    - items
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.OrderHistory:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      itemsJSON:
        type: string
      orderedAt:
        type: string
      originalOrderID:
        type: string
      sessionID:
        type: string
      status:
        type: string
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      table:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Table'
      tableID:
        type: integer
      tableNumber:
        type: integer
      totalAmount:
        type: integer
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.OrderItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      menu:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Menu'
      menuID:
        type: integer
      menuName:
        type: string
      order:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Order'
      orderID:
        type: string
      price:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: integer
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.OrderItemRequest:
    properties:
      menu_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - menu_id
    - quantity
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.OrderStatus:
    enum:
    - PENDING
    - CONFIRMED
    - PREPARING
    - COMPLETED
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusConfirmed
    - OrderStatusPreparing
    - OrderStatusCompleted
  github_com_kohwg_gowoopi_backend_internal_model.StatusUpdateRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.Store:
    properties:
      createdAt:
        type: string
      defaultLanguage:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.Table:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      isActive:
        type: boolean
      passwordHash:
        type: string
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      tableNumber:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.TableSession:
    properties:
      createdAt:
        type: string
      endedAt:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      startedAt:
        type: string
      store:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Store'
      storeID:
        type: string
      table:
        $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Table'
      tableID:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_kohwg_gowoopi_backend_internal_model.TableSetupRequest:
    properties:
      password:
        type: string
      table_number:
        type: integer
    required:
    - password
    - table_number
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 테이블오더 서비스 API
  title: Gowoopi Table Order API
  version: "1.0"
paths:
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: 매장 관리자 로그인
      parameters:
      - description: 로그인 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      summary: 관리자 로그인
      tags:
      - Auth
  /api/admin/menus:
    post:
      consumes:
      - application/json
      description: 새 메뉴 추가 (관리자)
      parameters:
      - description: 메뉴 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Menu'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 메뉴 생성
      tags:
      - Menu
  /api/admin/menus/{id}:
    delete:
      description: 메뉴 삭제 (관리자)
      parameters:
      - description: 메뉴 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 메뉴 삭제
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: 메뉴 정보 수정 (관리자)
      parameters:
      - description: 메뉴 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 수정할 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Menu'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 메뉴 수정
      tags:
      - Menu
  /api/admin/menus/order:
    patch:
      consumes:
      - application/json
      description: 메뉴 정렬 순서 변경 (관리자)
      parameters:
      - description: 순서 정보
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.MenuOrderRequest'
          type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 메뉴 순서 변경
      tags:
      - Menu
  /api/admin/orders:
    get:
      description: 매장의 전체 주문 목록 (관리자)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Order'
            type: array
      security:
      - BearerAuth: []
      summary: 매장 주문 조회
      tags:
      - Order
  /api/admin/orders/{id}:
    delete:
      description: 주문 삭제 (관리자)
      parameters:
      - description: 주문 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 주문 삭제
      tags:
      - Order
  /api/admin/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: 주문 상태 업데이트 (관리자)
      parameters:
      - description: 주문 ID
        in: path
        name: id
        required: true
        type: string
      - description: 상태 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.StatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 주문 상태 변경
      tags:
      - Order
  /api/admin/tables/{id}/complete:
    post:
      description: 테이블 세션 종료 및 주문 히스토리 이동 (관리자)
      parameters:
      - description: 테이블 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 테이블 이용 완료
      tags:
      - Table
  /api/admin/tables/{id}/history:
    get:
      description: 테이블의 과거 주문 내역 조회 (관리자)
      parameters:
      - description: 테이블 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 시작일 (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: 종료일 (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.OrderHistory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 테이블 주문 히스토리
      tags:
      - Table
  /api/admin/tables/setup:
    post:
      consumes:
      - application/json
      description: 새 테이블 생성 (관리자)
      parameters:
      - description: 테이블 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.TableSetupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Table'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 테이블 설정
      tags:
      - Table
  /api/auth/refresh:
    post:
      description: Access token 갱신
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 토큰 갱신
      tags:
      - Auth
  /api/customer/login:
    post:
      consumes:
      - application/json
      description: QR코드 스캔 후 테이블 세션 생성
      parameters:
      - description: 로그인 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      summary: 고객 로그인
      tags:
      - Auth
  /api/customer/menus:
    get:
      description: 매장의 카테고리별 메뉴 목록
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 메뉴 목록 조회
      tags:
      - Menu
  /api/customer/orders:
    get:
      description: 현재 세션의 주문 목록 (고객)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Order'
            type: array
      security:
      - BearerAuth: []
      summary: 내 주문 조회
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: 새 주문 생성 (고객)
      parameters:
      - description: 주문 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kohwg_gowoopi_backend_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 주문 생성
      tags:
      - Order
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"